+++ packages/react-testing/src/compat.ts
@@ -1,3 +1,6 @@
+import ReactDOM from 'react-dom';
+import type {Root as ReactRoot} from 'react-dom/client';
+
 import {ReactInstance, Fiber} from './types';
 
 export function getInternals(instance: ReactInstance): Fiber {
@@ -8,4 +11,26 @@ export function getInternals(instance: ReactInstance): Fiber {
   }
 
   return instance._reactInternals;
+}
+
+/** Shim to provide createRoot backwards compatibility for React < 18 */
+function createRootShim(element: HTMLElement): ReactRoot {
+  return {
+    render(children: React.ReactChild | Iterable<React.ReactNode>): void {
+      ReactDOM.render(children as any, element as any);
+    },
+    unmount(): void {
+      ReactDOM.unmountComponentAtNode(element as any);
+    },
+  };
+}
+
+/** Uses React >= 18 createRoot if available or falls back to shim*/
+export function createRoot(element: HTMLElement): ReactRoot {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    return require('react-dom/client').createRoot(element);
+  } catch {
+    return createRootShim(element);
+  }
 }